Added Gradiente_CT-80E, 80 columns expansion cart
Thanks to Maurício (iamweasel) for providing the initial hardware config file
and the ROM dump. Thanks to BiFi for figuring out why this initial config
didn't work.

This cartridge contains a V9938 chip. So if you use it you have a 2nd
videosource (use 'set videosource "Gradiente CT-80E"' or just 'cycle
videosource' to switch to it). OpenMSX already supported a secondary V9938.
What was not yet supported (and what BiFi figured out) is that the IRQ pin of
that V9938 is not connected to the MSX cartridge slot.

So basically all what this patch does is adding support for a not-connected IRQ
pin in the VDP code. This is done via the new 'OptionalIRQHelper' class. This
class will check the hardware configuration file for the new (boolean)
<irq_connected> tag (when not specified it defaults to 'true').

This patch is relatively large because it also removes the unused 'NMISource'
and 'DynamicSource' helper classes from the code (I kept them mostly as
examples, but now that we have OptionalIRQ that's no longer needed). I also had
to move some code from the .cc to the .hh file, this also makes the patch
larger.