Simplify idle channel detection in YM2413Okazaki
The original YM2413Okazaki code would calculate the sound output sample by
sample, and for each sample calculate channel by channel. IOW the outer loop
goes over all samples, and the inner loop goes over all channels. Quite a while
ago already (revision 9614a1c4a, Dec 2007) I swapped this order: outer loop
goes over channels, inner loop over samples. Over time this enabled many
optimizations because within one channel certain properties remain constant (as
long as the register values don't change) and that allows to create much
smaller specialized inner loops that have almost all conditionals removed.
(Very roughly: the new code is about 2x as fast as the old).

Another (much older) optimization we added to the YM2413Okazaki core was to
detect channels that are idle (=silent) and skip calculating output for them.
At the start of the generateChannels() method we would calculate one bit per
channel that indicates whether it's idle or not. Then before doing the actual
computations for that channel we check the corresponding bit. This is a useful
approach when that bit can be checked many times (so with the original order of
inner/outer loops). But when we only test it once (with outer loop going over
the channels), we don't gain anything anymore.

So this patch removes the initial phase of idle-channel detection. Instead it's
now integrated in the (outer) loop that goes over all channels. This allowed to
get rid of the 'idleSamples' stuff. And it has the additional bonus that
'am_phase' and 'pm_phase' are also updated on a completely idle YM2413 (there's
no benefit anymore in not updating them).