[3/5] Made destructors non-virtual where possible
The -Wsuggest-final-types warning (explained earlier in this patch series)
triggered _a_lot_ of times, much more than I expected. I didn't investigate in
detail, but I suspect a large portion of those is because of virtual
destructors. Instead of relying on de-virtualization, it's better to not have
virtual calls in the first place (C++ is not Java ;-). And it turned out that
in quite a few places we could indeed make our destructors non-virtual. That's
what this patch does.

Some background info:

Forgetting to make a destructor virtual can be a real source a bugs. Whenever
you delete an object via a pointer to (one of) its base class(es) and you
forgot to make the destructor of that base class virtual, you have undefined
behavior according to the c++ standard. In practice this undefined behavior
often manifests itself as resource leaks (because the members of the subclass
aren't destroyed) or even as heap or memory-pool corruption (because you free
an object of the wrong size).

To help protect against this type of bugs older compilers often implemented
a warning along these lines:
 - when a class has a virtual method, its destructor must also be virtual
Often this rule is too strict (and sometimes it's not strict enough). But to
silence the compiler warning we made the destructor virtual anyway, and that's
why we had too many virtual destructors in our (already reasonably old) code
base. Luckily compiler warnings have improved: with more recent compilers you
only get a warning when you do a polymorphic delete on a class with non-virtual
destructor.

These are the (new) generally recommended guidelines to decide between virtual
or non-virtual destructors:
- For base classes:
  * prefer to make the destructor *non-virtual* and *protected* (this is
    efficient and prevents accidental polymorphic delete)
  * but if you do need polymorphic delete, make the destructor *public*
    and *virtual*
- For non-base classes (final classes) you don't need virtual destructors

In an earlier patch in this series we added 'final' annotations on many (most)
classes. So the above rule now becomes very easy to follow:
 - class is final(*) -> destructor non-virtual(**)
 - not final, destructor should be either(***)
    * protected and non-virtual
    * public and virtual
(*) It's highly recommended to actually mark the class final. This prevents
    someone accidentally inherits from it in the future (and that's a bug
    because the destructor is non-virtual).
(**) It's possible the destructor is still virtual because one of the base
     classes has a virtual destructor, but we don't repeat the virtual keyword.
(***) So for base classes the default destructor is wrong (it's public and
      non-virtual).

After this patch the size of the generated code became 54kB smaller! It's
likely also a bit faster, but we don't have object deletions in performance
critical code paths, so I doubt you can measure this speed improvement.