[4/5] Mark virtual method overrides with new c++11 'override' keyword
(Strictly speaking 'override' isn't a c++ keyword, but that's a detail).

Copied from https://stackoverflow.com/questions/18198314/override-keyword-in-c
  'override' serves two purposes:
  - It shows the reader of the code that this function is a virtual function
    that is overriding a function in the base class.
  - The compiler also knows that it's an override, so can "check" that you are
    not altering/adding new functions that you think are overrides.

This patch uses 'override' wherever possible. I used 'clang-modernize' to
generate these (mechanical) changes. But I did manually check all changes. In
fact I found two bugs while doing this: in NowindInterface and in MegaSCSI the
peekMem() method name was misspelled (respectively as peek() and peekmem()).
Also fixed that in this patch. It _may_ also have found a bug in the
DiskCommand (it inherits from Command instead of RecordedCommand, so the
needRecord() method it implements is not an override), though I need more time
to investigate this. (If required) I'll fix it in another patch.

Before this patch, when overriding a virtual method, we used to repeat the
virtual keyword in the subclass. This repetition is not needed (code means
exactly the same with or without), but it was a useful reminder for the reader
of the code. Now that we can mark overrides in a more explicit way I decided to
leave out the repeated virtual keyword. Using google you find about equally
many examples that do or do not repeat virtual in combination with override,
we'll see in a few years what the generally used coding style becomes.