Fixed bug in allowTooFastAccess mode
Reported by sd_snatcher (post of 04-10-2014, 00:35) here:
  http://www.msx.org/forum/msx-talk/openmsx/wanted-feedback-mac-os-x-users?page=14

Background info:
 When the Z80 reads a byte from VRAM (via a read from IO port #98), the VDP
 cannot always immediately perform that read (because the bandwidth to the same
 RAM chips is shared among various different VDP subsystems, and you have to
 wait till a CPU access slot comes along). However to not have to stall the Z80
 IN instruction we need to immediately return a result. To solve this problem,
 the VDP always reads (and buffers) one byte in advance. So when you set the
 VRAM read address (via two OUT's to port #99) the VDP already reads (and
 buffers) the first byte. When you later perform the actual IN instruction, you
 get that buffered byte and the VDP schedules reading the next byte from VRAM,
 and so on.

Now for the bug: the real VDP only has one byte storage to buffer the result,
and that's also how it was implemented in openMSX. A read from VRAM would
 - schedule the next read from VRAM
 - return the previously buffered byte
Though allowTooFastAccess mode was recently changed to not actually schedule
that read anymore but immediately execute it (see revision 3ab96c51e). But this
already overwrites the buffered byte and thus we returned the wrong value.
Easily fixed once the problem was identified.