[1/3] Remove pimpl from MSXMotherBoard
Pimpl as in 'pointer to implementation', not as in pimple ;-)

The pimpl-idiom is useful as a compilation firewall: changes in the private
parts of the pimpl class don't trigger recompilations in users of that class.
So this results in faster incremental compilations.

This patch removes the pimpl-idiom from the MSXMotherBoard class. It's a
preparation for the next patch that will add a templatized method to
MSXMotherBoard. I believe it should be possible to add a templatized method to
a pimpl-class, though it is more complex. And the pimpl stuff is anyway less
useful nowadays (for a small project like openMSX):
 - Changes in MSXMotherBoard are less frequent than in the past
 - Compiler have gotten faster
 - Computers have gotten faster (faster cores, but especially more cores per
   computer).
Also this patch allows to simplify the MSXMotherBoard code quite a bit, and it
even is a tiny bit faster now.