Fixed detection of Philips Music Module in T2SE
To detect the presence of MSX-AUDIO, Tetris 2 Special Edition (T2SE) did
'INP(4)' and expected to see the value 0 (thanks to BiFi for figuring this
out). IO port 4 contains the Philips Music Module ACIA chip (used for MIDI).
(So this is probably not the best routine to check for MSX-AUDIO).

Our current implementation of that ACIA chip (a MC6850) is very basic. Before
this patch it returned a fixed value '2' on reading from port 4. This value 2
was added in revision 15403bb4c8 to make the detection routine in 'FAC demo 5'
pass.

In this patch I _slightly_ extend the implementation of the MC6850 chip so that
both detection routines pass. See comments in patch for more details.