[2/3] Cleanup MSXMotherBoard::getSharedStuff()
Before this patch MSXMotherBoard::getSharedStuff() was a big hack. It was
basically a reference-counted void-pointer where each user of this method had
to reimplement the reference-counting semantics.

Now that we switched to c++11 we have weak_ptr (this is sort-of like a
shared_ptr that doesn't contribute to the reference count) and variadic
templates. The combination of these two features allows to implement
getSharedStuff() in a much cleaner way.

'getSharedStuff()' still uses reference-counted void-pointers under the hood.
But there is only a single implementation anymore, and much of the
implementation details are handled by the std components itself
(std::shared_ptr and std::weak_ptr).

You often hear in presentations that modern c++11 (or c++14) code should rarely
require the 'new' and 'delete' keywords anymore. This patch makes another big
step in that direction.