Fix bug#539 KANA/CODE key not working in CHARACTER mode
This was a regression introduced in 259fc71b. That patch changed an internal
structure in EnumSetting from map<string,T> to a sorted vector<pair<string,T>>,
with T some enum type. This structure represents the mapping between the user
visible setting string and the internally used enum value. The (internal) base
class has an equivalent structure with 'T' replaced by 'int'. EnumSetting takes
such a structure as one of its constructor parameters. This data is actually
stored in the base class, so at some point we convert 'T' to 'int'. To avoid
copying all the string objects they are moved instead. The move-vs-copy change
was also part of patch 259fc71b.

Now the problem: the KeyboardSettings class defines 4 enum settings
(kbd_code_kana_host_key, kbd_deadkey1_host_key, kbd_deadkey2_host_key and
kbd_deadkey3_host_key). They all have the same string<->enum mapping. The code
(obviously) constructs this mapping structure only once and passes it to the 4
constructors. Though because internally the constructor moves from this
structure, the structure changes and the last 3 settings get created with an
invalid mapping (this in undefined behavior according to the c++ standard, in
practice it meant that all enum values were associated with empty strings).

So the move-vs-copy change introduced the problem, one possible fix is to
revert that. On the other hand it's a useful optimization for the vast majority
of the enum settings. An alternative fix is to pass 4 distinct copies of the
mapping structure in the KeyboardSettings code. Though that makes the
EnumSetting code very fragile, I mean this makes it very easy to use this class
in a wrong way, thus reintroduce similar problems in the future. Luckily there
exists a fix that gives us the best of both:

Before this patch, the EnumSetting class took a 'Map&' constructor parameter (a
non-const reference to this mapping structure). The fact that it's non-const
already hints that something weird is going on. If we choose the 1st solution,
we should make the type of this parameter 'const Map&' and then the
implementation of EnumSetting cannot move from it. Instead this patch makes the
type 'Map&&', an r-value reference. So this makes it clear to callers that the
parameter will be 'consumed' by EnumSetting (and thus that it's not allowed to
pass it to multiple constructors). And most important: the compiler will
enforce this, so the code in KeyboardSettings would not have compiled after
patch 259fc71b (and silently give wrong results).

Because the compiler now enforces we only pass r-values, we had to change some
of the callers. In most cases that was actually a cleanup (less temporary
variables). (Only) in case of the KeyboardSettings code it required taking a
copy for (3 of the 4) calls.