Replace our 'noncopyable' helper with c++11 '=delete'
Sometimes objects of a class shouldn't be copied for correctness reasons. C++
allows you to make the class non-copyable. So even if the current code doesn't
make such invalid copies, it prevents accidentally adding them in the future.

Pre-c++11 the idiom to make a class non-copyable was to give it a private
definition (but no implementation) of the copy-constructor and copy-assignment
operator. We abstracted this a bit by creating a helper base class (called
'noncopyable') with these properties and inheriting from that class.

In c++11 the preferred idiom is to give the class a 'deleted' copy constructor
and assignment operator (one reason for this is that it gives better error
messages, possibly already at compile-time instead of link-time). Also in c++11
many more classes are automatically made non-copyable, e.g. because they
contain move-only member variables (typical example is unique_ptr).

So this patch replaces the old with the new idiom. Though when non-copyability
is now implied I sometimes just removed the annotation. And in a few cases I
made a class now non-copyable where before we forgot to do this.