Fix clang compiler warnings
Compiled with clang-3.5 and clang-3.6 (pre-release) and fixed all compiler
warnings. The warnings in WD2793 and Y8950 were harmless. The stuff in
LaserdiscPlayer was a theoretical problem, but I don't think it could cause
problems in practice ('abs(<some-int64-type>)' could call the old c-library
function that takes an int (32-bit) parameter, but std::abs() has overloads to
handle 64-bit).