Implemented "too_fast_vram_access_callback"
Called when the CPU reads or writes VRAM (IO port #98) when the previous
request has not yet been handled. This can for example be used to break the
emulation so that it's possible to figure out exactly which Z80 instruction
triggered this condition and exactly where in the frame the VDP is executing
(e.g. border or display area).

For example:
  proc warn_too_fast {} {
    message "Too fast VRAM access PC=0x[format %x [reg PC]] VDP-line=[machine_info line_in_frame]" warning
  }
  set too_fast_vram_access_callback warn_too_fast

I've added similar examples to the 'umrcallback.tcl' script. This callback is
not enabled by default.