Fixed corruption in (some) XSA disk images
Regression introduced in revision 74a1f4cfa. The problem was that a manual
memory copy loop was replaced with a call to memmove() but both have different
semantics when the source and destination memory ranges overlap.

More in detail: I was aware that using memcpy() results in undefined behavior
when the source and destination buffers overlap. The manual page of memcpy()
explicitly points to memmove() for that case. Though memmove acts as-if the
source buffer is first copied to temporary storage and then copied to the
destination (practical implementations do either a forward or a backward copy).
This avoids mirroring parts of the input buffer multiple times in the
destination buffer. Though in case of a LZ77-like decompresser this mirroring
effect is exactly what you want. So neither memcpy() nor memmove() is a correct
replacement for the original loop.

Once the problem was identified the fix is easy: again do the copy using a
manual loop similar to the original one (can't be the same as the original code
because the buffer management was totally changed, and that was also the reason
why this copy routine was changed in revision 74a1f4cfa).