Removed PRT_DEBUG
A debug flavour build was very verbose. For the last 5+ years, I can't remember
when this output was actually useful for me. And I do remember many occasions
where I had to disable the debug output because it was too verbose or because
it interfered with more specific debug output I had added to investigate a
particular problem. Thus this patch removes all PRT_DEBUG debug output.

I took care to not remove useful information from the source code. So I removed
PRT_DEBUG statements that only repeat the name of the function, or the name of
an enum value in a switch-statement, or print values of input parameters, or
the return value. But debug prints that contain information that cannot be
deduced from the line(s) directly above or below it are only converted into
comments.