Partly revert previous commit
I mistakenly thought visual studio offered 'max_align_t' in both the global
namespace as in the std namespace. Instead it's only present in std (and that's
how it should be according to the c++ standard). That means the workaround for
gcc-4.7 and gcc-4.8 doesn't work. And we'll have to wait till gcc-4.9 before we
can remove the current workaround.