Added fast_stou() routine
I noticed that sometimes we would:
- construct a std::string object
- (only) convert that to an integer
- destroy the object again

This seems a bit inefficient, though we can't (always) use the libc functions
like strtol() because the string is not necessarily zero-terminated (or may not
even have any readable character after the last digit).

Often the strtol() like functions do too much work, e.g. they will strip
leading whitespace or leading +/- signs, or handle different bases, ...

So this patch adds a very simple (but also limited in functionality)
fast_stou() routine, that does conversion of a decimal string into an unsigned
integer. See comments in the code for the full list of limitations.

This patch only adds the routine, but doesn't use it yet.