Simplify CartridgeSlotManager::getSlotNum()
Before this patch the code would do a string->int conversion and then check
whether the result was in the range [0..3]. Though such single-digit integers
can be parsed in a much simpler way (even simpler than using the fast_stou()
routine introduced in the previous patch).

At the same time I restructured the if-chain in this routine to do less
redundant checks.