Return TclObject (iso string) when executing Tcl
Changed the return type of the methods
 - CommandController::executeCommand()   (and subclasses)
 - Interpreter::execute()
 - Interpreter::executeFile()
 - TclObject::executeCommand()
from std::string to TclObject. The main motivation for this change was the
following scenario:
 - Call an internal Tcl proc like openmsx::tabcompletion or
   openmsx::all_command_names.
 - This would (in a Tcl script) construct a list. The string representation of
   that list is returned to the calling c++ code.
 - The c++ code would then split that string again into elements and do
   something with those elements.
Now that we return a TclObject, the join/split operation can be avoided. In
other scenarios we sometimes avoid string<->int conversions and dynamic memory
(de)allocations to manage the internal storage of std::string objects.

TclObject is extended with a few methods to make it easier to treat it like a
list of strings. And Completer::completeString() is extended to (also) take a
pair of iterators to denote a range of possible values. Combined these two
extensions allow to directly use TclObjects for operations like tab-completion
(I mean without constructing extra intermediate temporary data structures).

After the above changes the Interpreter::splitList() methods are no longer
used, so they are removed.