[1/9] unique_ptr->value, setting objects
In this patch series we replace:
	const std::unique_ptr<MyClass> myMemberVariable;
with:
	MyClass myMemberVariable;
So instead of allocating a separate MyClass object on the heap and storing a
pointer to it, we now directly embed that object in the surrounding object.

This refactoring has one big disadvantage: before a forward declaration of
MyClass was sufficient in the header file, now we need the full declaration (so
replace 'class MyClass' with '#include "MyClass.hh"'). This results in more
include dependencies and thus on average more work on incremental compilations.
Though compilers and CPUs (multi-core) have gotten faster over the years so
for small to medium sized projects like openMSX this is less of an issue.

The refactoring has also some advantages:
- Possibly most important, the code becomes simpler: for the full patch series
  we have
    272 files changed, 3892 insertions(+), 4890 deletions(-)
  so netto this removes ~1000 LOC
- Less memory (de)allocations resulting in (slightly) faster code. Also less
  pointer dereferences and better cache locality. I measured about 1% faster
  startup speed. And about 200 less heap-(de)allocations. The resulting openMSX
  binary is approx 4kB smaller.
- Less wasted memory: there's (only) 4kB less heap in use. About 1kB of this
  comes from less pointers (1 pointer is 8 bytes on my system), and 3kB comes
  from less wasted heap (fragmentation + over-allocation).

I semi-arbitrarily split this series in several patches. This patch mostly
replaces setting objects.