Round instead of truncate inside DynamicClock::setFreq()
Internally openMSX works with a time resolution of 3579545*960Hz. All clocks
tick at an integer multiple of this master clock. This means that not all
frequencies can be represented exactly. The DynamicClock class represents a
clock whose rate can be changed at runtime via the setFreq() method. The
implementation of setFreq() divides the master clock rate with the requested
clock rate. Before this patch, the result of that division was truncated, now
it's rounded to the nearest integer. Thus if the requested clock rate could not
be represented exactly we now take the closest (slightly) higher or lower
frequency, before we always took the closest higher frequency.

This change is motivated by bug
   #559 PSG/Z80 timing combination shows something is off
   http://sourceforge.net/p/openmsx/bugs/559/

The PSG produces sound at a frequency of 3579545/16Hz (~223722Hz). That is a
clock rate that _can_ be represented exactly in openMSX, namely one that is
exactly 960*16=15360 times slower than the master clock. Before this patch, due
to a combination of rounding and truncation we actually ended up with a clock
that's 15359 times slower than the master clock (or ~223736Hz). The error is
less than 0.006%, but still this tiny difference was enough to cause the slight
modulation described in the above bug report.

Side note:
Because we know the ratio of desired-clock and master-clock should be exactly
15360, it seems like a detour (with risk of rounding errors) to calculate this
ratio at runtime. Unfortunately this value 15360 is only correct when we're
emulating at 100% real-time speed. If the user sets e.g. speed=80 or speed=120
we actually do need to recalculate that value.