Fixed (rare) clipping bug in sprite-mode-1 drawing
To reproduce run "Battle Ship Clapton 2" on a turbor machine. Simply leave the
'demo' running, and after a while (several minutes) more and more pixels in the
right border will get corrupted.

The SpriteConverter::drawMode1() takes a minX and maxX parameter, these
determine a range (in a horizontal line) where the sprites should be drawn. (We
don't always draw complete lines from the left to the right border). When the
sprite pattern starts before minX, we need to clip the pattern. Similarly we
need to clip when the pattern extends past maxX. In sprite mode 1 the code had
a bug when we needed to clip at both minX and maxX: it would (wrongly) only
clip at minX.

In sprite mode 2 we need to do the exact same clipping. That code was correct.
So in this patch I factored out the (correct) clipping code and use it now for
both modes.

As far as I can tell this bug has always been present in openMSX. Though only
after revision 580607b (2014/06/25) pixel errors in the left/right border
accumulate (we don't redraw the border anymore if it doesn't change). So before
bugs like this were only visible for a single frame, and thus we didn't notice
them.