Replace 'typedef' with 'using'
C++11 allows to replace
	typedef vector<Foo> MyVector;
with
	using MyVector = vector<Foo>;
The latter has more natural left-to-right syntax (similar to 'int a = 3').
Though the real advantage is that it allows for templatized typedefs. (Not
possible with the old syntax, so we didn't use those yet). See here for a bit
more background info:
  https://stackoverflow.com/questions/10747810/what-is-the-difference-between-typedef-and-using-in-c11

I think vs2013 already supports this c++11 feature, but if not we can easily
revert this patch.