Fix bug#566: Extension AVT_DPF-550 crashes openMSX
The problem seemed to be that the software instructed the WD2793 to read a
sector while the disk drive wasn't spinning. When the WD2793 doesn't find the
requested sector within 5 revolutions it will generate a RECORD NOT FOUND
error. Though when the drive isn't spinning that condition doesn't occur
anytime soon. So the DiskDrive::getTimeTillIndexPulse() method calculated that
time as 'infinity'. But the Scheduler::setSyncPoint() method didn't like
scheduling events that only occur at infinity.

Fixed by letting the WD2793 immediately generate an error in this condition.
How does the real WD2793 behave in this situation? I also added an assert in
Scheduler that should make diagnosing this type of mistakes easier in the
future.