Fixed ROM access to RS-232C ROM.
The old code assumed it was 16kB. If it was smaller, it would crash. If
it was larger, it would be ignored. Fixed by mirroring on its size and
checking that the ROM is 8kB or 16kB, because other sizes are not
supported by the current implementation (and are not known so far).

This will cause 'overdumped' ROMs to stop working. There was at least one
of such, which I will fix in the next commit.