Create new snapshots when fast-forwarding from a distant snapshot
ReverseManager::goTo() will fast-forward to the requested time from the
most recent snapshot before it. However, in some cases "most recent" is
a long time ago, for example after loading a replay, since those have
relatively few snapshots in them. While we have to re-emulate a lot of
frames the first time we reverse into a sparsely populated snapshot
area, we now take snapshots during that first time to make subsequent
reverse actions much faster.

The scheduling of the next snapshot was moved out of the takeSnapshot()
method since we want to be able to take a snapshot without scheduling
another.