Revert "Simplify computation of preTarget in ReverseManager::goTo()"
This reverts commit 738d7928b3f77ab391a0073ba83713e1fbe083e9.

Logically the simplification of this reverted commit makes sense. Though
EmuTime is always 'unsigned', and subtracting an EmuDuration from an EmuTime
will assert that the result is not a negative (unrepresentable) EmuTime.

This assert actually triggered when I was loading an old savestate (when
testing the fix from commit 9de5c5af8d).