Don't make snapshots every second while fast forwarding
But make them always half-way the remaining period, with a minimum
distance of the usual interval. This massively speeds up the
fast-forwarding and practically none of the snapshots created while
doing this are pruned. And it leaves exactly the snapshots you need near
the end.

This is an improvement on 6fb196c6a2b3a2aa3e6ed9e9e0cb950104d58d21.
I have a very large test replay where I jumped to 116497, while the last
snapshot before that was at 104704, so a fast-forward of 11794 seconds
needed to be made.
Before the original patch, this took about 31 seconds, but after the
patch it took about 57 seconds. With this new patch, it takes about 32
seconds. 14 extra snapshots were taken in total.