Improve implementation of auto-save replay
Now we use a proper setting and put a trace on it. You can now save this
setting and auto-save will be enabled when starting up again. (This was
the main motivation.)
Note that saving will cause a (small) stutter, because it will take
time. I'll try to fix that in a later patch by reducing the amount of
snapshots in the replay.