Use c++11 functions log2() and exp2()
Replace
   log(x) / log(2)  ->  log2(x)
   pow(2, x)        ->  exp2(x)
The functions log2() and exp2() are new in c++11 (copied from c99). They more
directly specify the intent. (They may or may not execute faster and they may
or may not yield a more numerically accurate result, but neither matters much
in this context).

Note that there is a log10() function in c++11 (we're already using it) but no
exp10() function (it's part of glibc as an extension, it's not part of c++11).

I've also simplified
  log(1 / x)  ->  -log(x)