[1/5] Remove pimpl from Y8950
When building openmsx with gcc-5.1 using LTO (Link Time Optimizations) I got a
warning about ODR-violations (One Definition Rule). And indeed both Y8950.cc
and YMF262.cc instantiated enum_string<EnvelopeState> with each a different
definition for the EnvelopeState enum.

Easiest fix is to qualify both EnvelopeState enums, so either define them in
the Y8950/YMF262 class or in some (different) namespace. I initially tried the
namespace approach, but it was more difficult than it should be because of the
pimpl-idiom that was used by both classes.

In lots of other code we've already removed the pimpl-idiom because compile
times are no longer a big issue (no longer worth the added complexity). As far as
I can see there are only 5 classes remaining that use pimpl:
  Y8950, YMF262, YMF278, YM2151 and VLM5030

This patch series removes the pimpl-idiom from these 5 classes. This moves the
EnvelopeState enum into the Y8950/YMF262 class and thus solves the
ODR-violation.