Optimize (colored) text drawing in the console in openGL
The openmsx console rendering code keeps a cache of previously rendered text
fragments (because mostly the same text is drawn on consecutive frames). The
console also supports colored text.

Before this patch two fragments of the same text but with a different color
were distinct cache objects. For the SDL renderer this is still the case after
this patch. For the openGL renderer however we now always render text in white
(to a texture) and apply the color only in the actual draw command.

The advantage is that if the same text fragment appears in multiple colors in
the console (in openGL) we only have to render that text once (and draw the
texture multiple times in different colors). Drawing in different colors
(implemented as multiplying the texture colors) is practically free in openGL.
In fact the current shader already contained those color multiplications (but
all but the alpha-component were multiplications by 1.0 before).