Store Breakpoints by value
Before we stored breakpoints as
   vector<shared_ptr<BreakPoint>>
now they're stored as
  vector<BreakPoint>

When one (or more) breakpoint(s) hit, we execute a Tcl command. It's possible
that Tcl command will add or remove other breakpoints (= change the vector
we're iterating over) or remove the current breakpoint (= delete the BreakPoint
object we're currently using). To avoid these two potential problems, before
starting to execute the Tcl commands, we take a copy of the subset of the
to-be-executed breakpoints. This way we can freely modify the breakpoint vector
(we're now iterating over the copy) and even delete the current breakpoint
object (we're working on the copy).

Because we're making copies, the original idea was to store the BreakPoint
objects by shared_ptr, that way copying an object can be done by increasing a
reference count. Now instead we make an actual copy of the BreakPoint objects.
Let's compare both approaches:

- Typically we frequently (every instruction(*)) check whether a breakpoint is
  hit, but only rarely execute a breakpoint. So we should optimize for the
  former case. Checking whether a bp is hit requires checking the address
  fields of the bp objects (the bp's are sorted on address, so this is a
  O(log(n)) operation). If we store the objects by (shared) pointer we have an
  extra indirection to access the address field.

- Breakpoint objects are relatively small (about 24 bytes) so copying them is
  anyway very cheap. Copying a shared_ptr requires incrementing a reference
  count with atomic instructions (even though here we're only using a single
  thread), that is likely slower.

- Note that before revision 3cfaaf14 (made a few hours ago), breakpoint objects
  had two additional reference member variables. So it was technically not even
  possible to copy them (even though the references actually always had the
  same value).

(*) If there are no breakpoints defined, we have an optimized code path that
doesn't check for breakpoints.