Made File objects moveable + default-constructable
Before this patch a File object always refers to an open file. If the code
only sometimes needed an open file it typically used a unique_ptr<File> and
when no file is currently needed it could use a nullptr.

The previous patch made the internal implementation of a File object a single
(unique_)pointer. So if we have unique_ptr<File> we actually have a pointer to
a pointer. That seems like an unnecessary indirection. Let's try to improve
that.

The main change of this patch is to give a File object an additional possible
state: it can now also represent a closed file handle. In other words a closed
file handle does not refer to a file (the internal representation of the File
object is a nullptr in this case). So now code can use a File object (by value)
to represent both open and closed files (closed is equivalent to no file in
this case).

Another reason why unique_ptr<File> was used is to pass File objects around.
The current implementation doesn't allow to make copies of File objects, so
passing unique_ptr's was a good fit. Though we can make File objects movable
and get rid of the extra indirection. This is the second major change of this
patch.

The rest of this patch (the majority) are mechanical changes that make use of
the two above features. So we change
- unique_ptr<File>   ->  File   (and becaue of this replace '->' with '.')
- check for nullptr  ->  check for is_open()
- file.reset()       ->  file.close()