Made Rom objects movable
OpenMSX has several rom mappers. They all contain a unique_ptr<Rom> instead of
holding the Rom object by value. This causes an extra indirection whenever the
content of the rom is being accessed (though the cacheline mechanism mitigates
most of this cost).

The reason for this extra indirection is that the Rom object is constructed
before the mapper object is constructed (for most other MSXDevices that
internally use a Rom this is not the case). The reason is that for mapper type
detection we first need to investigate the content of the rom before we know
the mapper type (e.g. by looking up the rom-checksum in a database or by some
mapper guessing heuristics).

The original solution was to first create a Rom object and then pass a
(unique_)pointer to this Rom to the mapper constructor. After this patch we
instead move the Rom object into the mapper object.

By far the majority of the changes in this patch are boring mechanical changes.
The only tricky part is the Rom move-constructor. Internally Rom contains a
RomDebuggable object, and that one has a back-pointer to the Rom object. So
when the Rom is moved that back-pointer needs to be updated.