Don't store size in MemBuffer
MemBuffer is basically a wrapper around malloc/realloc/free, though it's
type-safe and it offers extra functionality like alignment.

Before this patch MemBuffer internally stored both a pointer to the allocated
memory block and the size of that block. This patch removes that size field.

The size field was used to
- offer STL-style begin()/end() iterators, none of the openMSX code uses this
- offer a size() method
The size() method is used, but only by about half of the MemBuffer users. Those
users now have to manually keep track of the size.

This change makes MemBuffer really a zero-overhead replacement for malloc.