Remove FilePool/sha1 stuff from File class
The File class is a generic file handling class, it allows to read/write files,
it offers stuff like mmap(), it transparently handles (g)zipped files, etc. It
can also calculate the sha1 of a file in an efficient way by using the FilePool
cache.

Though this sha1 stuff adds a significant (relative) overhead to each File
object (the object size increases from 8 to 36 bytes). The interface of the
class and the class invariants also become uglier (e.g. the getSha1Sum() method
can only be called after setFilePool() is called, and setFilePool() can only be
called at most once). Only the minority of the users of the File class actually
need the extra sha1 functionality. So let's see if we can separate the basic
file handling from the extra sha1 functionality.

There was already a mechanism to calculate the sha1 for a file, namely
FilePool::getSha1Sum(File&). Under the hood File::getSha1Sum() was implemented
using the former function. Let's compare these two:

- FilePool::getSha1Sum(File&) does a (O(N)) lookup of the filename. If found
  and if the file-modification-timestamp matches we return a cached sha1 value.
  If not we read the whole file and calculate the sha1 (and store the result in
  the cache).

- File::getSha1Sum() has a local sha1-cache cache (O(1) lookup). So if the sha1
  of this File object has already been calculated before we can immediately
  return the result, otherwise we fall back to the previous method. Though only
  for certain file types we request the sha1 for the same File object more than
  once. More on this below.

  When the File::write() method is called, both the local cache in the File
  object and the FilePool cache for this file are invalidated. Though when a
  file is been written to, the file modification timestamp changes, and the
  FilePool cache already becomes invalid.

I mentioned that only for certain file types we repeatedly request the sha1.
These are the cassette, rom and disk images (because the sha1 ends up in the
period savestate snapshots of the reverse system).
- Actually for cassette images (wav or cas) the sha1 is only calculated once
  and cached in the CassetteImage object (ATM we do not support writing to wav
  files).
- Similar for rom images the sha1 is calculated once and cached in the ROM
  object. Detail: because of IPS patches the rom-sha1 might be different from
  the file-sha1. Also not all ROMs are backed by a file (see code for more
  details). That's why ROM has it's own sha1-cache instead of reusing the cache
  in the associated File object.
- For disk images the (old) code _mostly_ relies on the sha1-cache in the File
  object. Though because of stuff like IPS patches and DirAsDsk (not backed by
  a File object) also the Disk classes already had a local sha1-cache, it's
  just not used consistently (this patch changes that). Of course it is
  possible to write to disks, in that case the local sha1-cache must be
  invalidated.

So actually there's no good reason to keep the sha1 functionality in the
generic File class. And thus this patch removes it. As already mentioned, this
patch also reworks the sha1-caching mechanism in the Disk classes.